cleaning up tests.
